requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
ed25519==1.5
pydantic==2.6.1
typing-extensions==4.9.0
cairo-lang==0.12.0
eth-account==0.10.0
ledgereth==0.9.0
starknet-crypto-py==0.1.0
starknet.py==0.22.0
web3==6.11.3 